chore: separate controls and types
Change the @controls import and separate controls and types to their own import path
